controllers/user: resolve type errors with workaround
