+++ server/controllers/user.controller/index.ts
@@ -1,0 +1,2 @@
+export * from './apiKey';
+export * from './helpers';
